Translation/translationModel/requirements.txt
absl-py==0.8.1
astor==0.8.0
astroid==2.3.3
beautifulsoup4==4.7.1
bs4==0.0.1
cachetools==3.1.1
certifi==2018.11.29
chardet==3.0.4
Click==7.0
cycler==0.10.0
Flask==1.1.1
Flask-Cors==3.0.8
gast==0.2.2
google==2.0.2
google-auth==1.7.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.25.0
gunicorn==20.0.4
h5py==2.10.0
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mccabe==0.6.1
numpy==1.17.4
oauthlib==3.1.0
opt-einsum==3.1.0
protobuf==3.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycairo==1.18.1
PyGObject==3.32.1
pylint==2.4.4
pyparsing==2.4.5
python-dateutil==2.8.1
requests==2.21.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.21.3
scipy==1.3.2
six==1.13.0
sklearn==0.0
soupsieve==1.8
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
typed-ast==1.4.0
urllib3==1.24.1
Werkzeug==0.16.0
wrapt==1.11.2
